[Feat] 고객 게시판 리팩토링 및 ui 동작 일부 구현
